+++ R/validate_dynamical_system_inputs.R
@@ -7,6 +7,9 @@ validate_dynamical_system_inputs <- function(
     verbose = TRUE
 )
 {
+    # If the drivers input doesn't have a time column, add one
+    drivers <- add_time_to_weather_data(drivers)
+
     # The inputs to this function have the same requirements as the `run_biocro`
     # inputs with the same names
     error_messages <- check_run_biocro_inputs(
@@ -18,10 +21,7 @@ validate_dynamical_system_inputs <- function(
         verbose = verbose
     )
 
-    send_error_messages(error_messages)
-
-    # If the drivers input doesn't have a time column, add one
-    drivers <- add_time_to_weather_data(drivers)
+    stop_and_send_error_messages(error_messages)
 
     # Make module creators from the specified names and libraries
     direct_module_creators <- sapply(